+++ lib/yt/associations/has_reports.rb
@@ -3,6 +3,23 @@ module Yt
     # @private
     # Provides methods to access the analytics reports of a resource.
     module HasReports
+      # @!macro [new] reports
+      #   Returns the reports for the given metrics grouped by the given dimension.
+      #   @!method reports(options = {})
+      #   @param [Hash] options the metrics, time-range and dimensions for the reports.
+      #   @option options [Array<Symbol>] :only The metrics to generate reports
+      #     for.
+      #   @option options [Symbol] :by (:day) The dimension to collect metrics
+      #     by. Accepted values are: +:day+, +:month+.
+      #   @option options [#to_date] :since The first day of the time-range.
+      #     Also aliased as +:from+.
+      #   @option options [#to_date] :until The last day of the time-range.
+      #     Also aliased as +:to+.
+      #   @return [Hash<Symbol, Hash>] the reports for each metric specified.
+      #   @example Get the views and estimated minutes watched by day for last week:
+      #     resource.reports only: [:views, :estimated_minutes_watched] since: 1.week.ago, by: :day
+      #     # => {views: {Wed, 8 May 2014 => 12, Thu, 9 May 2014 => 34, …}, estimated_minutes_watched: {Wed, 8 May 2014 => 9.0, Thu, 9 May 2014 => 6.0, …}}
+
       # @!macro [new] report
       #   Returns the $1 grouped by the given dimension.
       #   @!method $1(options = {})
@@ -18,6 +35,11 @@ module Yt
       #   @example Get the $1 for each day of last week:
       #     resource.$1 since: 2.weeks.ago, until: 1.week.ago, by: :day
       #     # => {Wed, 8 May 2014 => 12.0, Thu, 9 May 2014 => 34.0, …}
+      #   @return [Hash<Range<Date, Date>, $2>] if grouped by month, the $1
+      #     for each month in the time-range.
+      #   @example Get the $1 for this and last month:
+      #     resource.$1 since: 1.month.ago, by: :month
+      #     # => {Wed, 01 Apr 2014..Thu, 30 Apr 2014 => 12.0, Fri, 01 May 2014..Sun, 31 May 2014 => 34.0, …}
       #   @macro report
 
       # @!macro [new] report_with_range
@@ -48,19 +70,19 @@ module Yt
 
       # @!macro [new] report_by_day
       #   @option options [Symbol] :by (:day) The dimension to collect $1 by.
-      #     Accepted values are: +:day+.
+      #     Accepted values are: +:day+, +:month+.
       #   @macro report_with_day
 
       # @!macro [new] report_by_day_and_country
       #   @option options [Symbol] :by (:day) The dimension to collect $1 by.
-      #     Accepted values are: +:day+, :+range+.
+      #     Accepted values are: +:day+, +:month+, :+range+.
       #   @macro report_with_day
       #   @macro report_with_range
       #   @macro report_with_country
 
       # @!macro [new] report_by_day_and_state
       #   @option options [Symbol] :by (:day) The dimension to collect $1 by.
-      #     Accepted values are: +:day+, :+range+.
+      #     Accepted values are: +:day+, +:month+, :+range+.
       #   @macro report_with_day
       #   @macro report_with_range
       #   @macro report_with_country_and_state
@@ -96,7 +118,7 @@ module Yt
 
       # @!macro [new] report_by_video_dimensions
       #   @option options [Symbol] :by (:day) The dimension to collect $1 by.
-      #     Accepted values are: +:day+, +:range+, +:traffic_source+,
+      #     Accepted values are: +:day+, +:month+, +:range+, +:traffic_source+,
       #     +:search_term+, +:playback_location+, +:related_video+,
       #     +:embedded_player_location+.
       #   @return [Hash<Symbol, $2>] if grouped by embedded player location,
@@ -122,7 +144,7 @@ module Yt
 
       # @!macro [new] report_by_channel_dimensions
       #   @option options [Symbol] :by (:day) The dimension to collect $1 by.
-      #     Accepted values are: +:day+, +:range+, +:traffic_source+,
+      #     Accepted values are: +:day+, +:month+, +:range+, +:traffic_source+,
       #     +:search_term+, +:playback_location+, +:related_video+, +:video+,
       #     +:playlist+, +:embedded_player_location+.
       #   @return [Hash<Symbol, $2>] if grouped by embedded player location,
@@ -135,7 +157,7 @@ module Yt
 
       # @!macro [new] report_by_playlist_dimensions
       #   @option options [Symbol] :by (:day) The dimension to collect $1 by.
-      #     Accepted values are: +:day+, +:range+, +:traffic_source+,
+      #     Accepted values are: +:day+, +:month+, +:range+, +:traffic_source+,
       #     +:playback_location+, +:related_video+, +:video+,
       #     +:playlist+.
       #   @macro report_with_channel_dimensions
@@ -179,8 +201,9 @@ module Yt
 
         define_metric_on_method metric
         define_metric_method metric
-        define_range_metric_method metric, type
-        define_all_metric_method metric
+        define_reports_method metric, type
+        define_range_metric_method metric
+        define_all_metric_method metric, type
       end
 
     private
@@ -191,20 +214,45 @@ module Yt
         end
       end
 
-      def define_metric_method(metric)
-        define_method metric do |options = {}|
-          from = options[:since] || options[:from] || (options[:by] == :day ? 5.days.ago : '2005-02-01')
+      def define_reports_method(metric, type)
+        (@metrics ||= {})[metric] = type
+        define_method :reports do |options = {}|
+          from = options[:since] || options[:from] || (options[:by].in?([:day, :month]) ? 5.days.ago : '2005-02-01')
           to = options[:until] || options[:to] || Date.today
           location = options[:in]
           country = location.is_a?(Hash) ? location[:country] : location
           state = location[:state] if location.is_a?(Hash)
-
-          range = Range.new *[from, to].map(&:to_date)
           dimension = options[:by] || (metric == :viewer_percentage ? :gender_age_group : :range)
+          if dimension == :month
+            from = from.to_date.beginning_of_month
+            to = to.to_date.beginning_of_month
+          end
+          date_range = Range.new *[from, to].map(&:to_date)
+
+          only = options.fetch :only, []
+          reports = Collections::Reports.of(self).tap do |reports|
+            reports.metrics =  self.class.instance_variable_get(:@metrics).select{|k, v| k.in? only}
+          end
+          reports.within date_range, country, state, dimension
+        end unless defined?(reports)
+      end
+
+      def define_metric_method(metric)
+        define_method metric do |options = {}|
+          from = options[:since] || options[:from] || (options[:by].in?([:day, :month]) ? 5.days.ago : '2005-02-01')
+          to = options[:until] || options[:to] || Date.today
+          location = options[:in]
+          country = location.is_a?(Hash) ? location[:country] : location
+          state = location[:state] if location.is_a?(Hash)
+          dimension = options[:by] || (metric == :viewer_percentage ? :gender_age_group : :range)
+          if dimension == :month
+            from = from.to_date.beginning_of_month
+            to = to.to_date.beginning_of_month
+          end
+          range = Range.new *[from, to].map(&:to_date)
 
           ivar = instance_variable_get "@#{metric}_#{dimension}_#{country}_#{state}"
           instance_variable_set "@#{metric}_#{dimension}_#{country}_#{state}", ivar || {}
-
           case dimension
           when :day
             Hash[*range.flat_map do |date|
@@ -216,23 +264,23 @@ module Yt
         end
       end
 
-      def define_range_metric_method(metric, type)
+      def define_range_metric_method(metric)
         define_method "range_#{metric}" do |date_range, dimension, country, state|
           ivar = instance_variable_get "@range_#{metric}_#{dimension}_#{country}_#{state}"
           instance_variable_set "@range_#{metric}_#{dimension}_#{country}_#{state}", ivar || {}
-          instance_variable_get("@range_#{metric}_#{dimension}_#{country}_#{state}")[date_range] ||= send("all_#{metric}").within date_range, country, state, dimension, type
+          instance_variable_get("@range_#{metric}_#{dimension}_#{country}_#{state}")[date_range] ||= send("all_#{metric}").within date_range, country, state, dimension
         end
         private "range_#{metric}"
       end
 
-      def define_all_metric_method(metric)
+      def define_all_metric_method(metric, type)
         define_method "all_#{metric}" do
           # @note Asking for the "earnings" metric of a day in which a channel
           # made 0 USD returns the wrong "nil". But adding to the request the
           # "estimatedMinutesWatched" metric returns the correct value 0.
-          query = metric
-          query = "estimatedMinutesWatched,#{metric}" if metric == :earnings
-          Collections::Reports.of(self).tap{|reports| reports.metric = query}
+          metrics = {metric => type}
+          metrics[:estimated_minutes_watched] = Integer if metric == :earnings
+          Collections::Reports.of(self).tap{|reports| reports.metrics = metrics}
         end
         private "all_#{metric}"
       end